+++ clap_complete/tests/snapshots/special_commands.zsh
@@ -36,7 +36,7 @@ _my-app() {
         case $line[3] in
             (test)
 _arguments "${_arguments_options[@]}" /
-'--case=[the case to test]: : ' /
+'*--case=[the case to test]: : ' /
 '-h[Print help information]' /
 '--help[Print help information]' /
 '-V[Print version information]' /
@@ -45,7 +45,7 @@ _arguments "${_arguments_options[@]}" /
 ;;
 (some_cmd)
 _arguments "${_arguments_options[@]}" /
-'--config=[the other case to test]: : ' /
+'*--config=[the other case to test]: : ' /
 '-h[Print help information]' /
 '--help[Print help information]' /
 '-V[Print version information]' /