+++ src/builder/command.rs
@@ -3789,10 +3789,6 @@ impl<'help> Command<'help> {
 
             let mut pos_counter = 1;
             let hide_pv = self.is_set(AppSettings::HidePossibleValues);
-            let auto_help =
-                !self.is_set(AppSettings::NoAutoHelp) && !self.is_disable_help_flag_set();
-            let auto_version =
-                !self.is_set(AppSettings::NoAutoVersion) && !self.is_disable_version_flag_set();
             for a in self.args.args_mut() {
                 // Fill in the groups
                 for g in &a.groups {
@@ -3815,24 +3811,6 @@ impl<'help> Command<'help> {
                     a.settings.set(ArgSettings::HidePossibleValues);
                 }
                 a._build();
-                // HACK: Setting up action at this level while auto-help / disable help flag is
-                // required.  Otherwise, most of this won't be needed because when we can break
-                // compat, actions will reign supreme (default to `Store`)
-                if a.action.is_none() {
-                    if a.get_id() == "help" && auto_help && !a.is_takes_value_set() {
-                        let action = super::ArgAction::Help;
-                        a.action = Some(action);
-                    } else if a.get_id() == "version" && auto_version && !a.is_takes_value_set() {
-                        let action = super::ArgAction::Version;
-                        a.action = Some(action);
-                    } else if a.is_takes_value_set() {
-                        let action = super::ArgAction::StoreValue;
-                        a.action = Some(action);
-                    } else {
-                        let action = super::ArgAction::IncOccurrence;
-                        a.action = Some(action);
-                    }
-                }
                 if a.is_positional() && a.index.is_none() {
                     a.index = Some(pos_counter);
                     pos_counter += 1;